Componentizes search dialog
