+++ app/frontend/lib/components/nav-main.svelte
@@ -3,57 +3,36 @@
   import { Link } from '@inertiajs/svelte'
   import { SearchSm } from '@voolt_technologies/untitledui-svelte'
   import { getOS } from '../utils'
+  import SearchDialog from './search-dialog.svelte';
   
-  import * as Command from "@/lib/components/ui/command/index.js";
-  import { File } from '@lucide/svelte'
   import type { DocumentType } from '@/pages/Document/types'
   import type { TeamType } from '@/pages/Team/types'
+
+  export type Item = {
+    title: string
+    url: string
+    tooltip: string
+    // This should be `Component` after @lucide/svelte updates types
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    icon: any
+    isActive?: boolean
+    badge?: number
+  }
 
   let {
     items, documents, teams
   }: {
-    items: {
-      title: string
-      url: string
-      tooltip: string
-      // This should be `Component` after @lucide/svelte updates types
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      icon: any
-      isActive?: boolean
-      badge?: number
-    }[],
+    items: Item[],
     documents: DocumentType[],
     teams: TeamType[],
   } = $props()
 
   const os = getOS()
 
-  console.log(documents);
-
   let isSearchDialogOpen = $state(false);
-  let filteredDocs = $state(documents);
-  let searchTerm = $state('');
 
   function openSearchDialog() {
     isSearchDialogOpen = true;
-  }
-
-  // TODO: Se inicialmente o Command.Dialog exibir TODOS os documentos,
-  // pode ser interessante, ao buscar por titulo, filtrá-los via JavScript
-  // ao invés de buscá-los via banco.
-  // Outras buscas (como por conteúdo) precisarão ser via banco mesmo.
-  function handleInput(event: Event) {
-    const input = event.target as HTMLInputElement;
-    searchTerm = input.value.trim();
-
-    if (!searchTerm) {
-      filteredDocs = documents;
-      return;
-    }
-
-    fetch(`/documents?search=${searchTerm}`, { method: 'GET' })
-      .then(res => res.json())
-      .then(res => filteredDocs = res);
   }
 </script>
 
@@ -82,37 +61,12 @@
     </Sidebar.MenuButton>
   </Sidebar.MenuItem>
 
-  <!-- Search dialog -->
-  <Command.Dialog bind:open={isSearchDialogOpen} shouldFilter={false}>
-    <Command.Input placeholder="Pesquise um documento ou seu conteúdo..." oninput={handleInput} />
-    <Command.List>
-      <Command.Empty>Nenhum resultado encontrado.</Command.Empty>
-      <Command.Group heading="Deste time">
-        {#each filteredDocs as doc (doc.id)}
-          <Command.Item>
-            <File class="size-5 text-muted-foreground" />
-            <span>{ doc.title || "Sem título" }</span>
-          </Command.Item>
-        {/each}
-      </Command.Group>
-      <Command.Separator />
-      <Command.Group heading="De outros times">
-        <Command.Item>
-          <File class="size-5 text-muted-foreground" />
-          <span>TODO</span>
-        </Command.Item>
-      </Command.Group>
-      <Command.Separator />
-      <Command.Group heading="Outras opções">
-        {#each items as item (item.title)}
-          <Command.Item>
-            <item.icon class="[&>path]:stroke-[2.5]" />
-            <span class="grow">{item.title}</span>
-          </Command.Item>
-        {/each}
-      </Command.Group>
-    </Command.List>
-  </Command.Dialog>
+  <SearchDialog
+    open={isSearchDialogOpen}
+    items={items}
+    documents={documents}
+    teams={teams}
+  />
 
   {#each items as item (item.title)}
     <Sidebar.MenuItem class="font-semibold">
